requirements.txt
selenium==4.15.2
python-dotenv==1.0.0
pandas==2.0.3
webdriver-manager==4.0.1
flask==2.3.3
flask-cors==4.0.0